adding auth
