+++ backend/auth_app/views.py
@@ -1,73 +1,111 @@
-import os
-import json
 import hashlib
 import hmac
-from django.contrib.auth.models import User
+import json
+import os
+import urllib.parse
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny
-from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework import status
+from .models import User as CustomUser
+from auth_app.serializers import UserSerializer
 
-
-def get_tokens_for_user(user):
-    refresh = RefreshToken.for_user(user)
-    return {
-        "refresh": str(refresh),
-        "access": str(refresh.access_token),
-    }
-
+# Получаем токен из переменной окружения
+TELEGRAM_BOT_TOKEN = os.getenv('TOKEN')
 
 class TelegramAuthView(APIView):
     permission_classes = [AllowAny]
 
     def post(self, request):
-        init_data = request.data.get("initData")  # Получаем initData
-        bot_token = os.getenv("TOKEN", "token")
+        init_data = request.data.get('init_data')  # данные из Telegram Web App
+        if not init_data:
+            return Response({"error": "init_data is required"}, status=400)
 
-        if not init_data or not bot_token:
-            return Response(
-                {"error": "Invalid data"}, status=status.HTTP_400_BAD_REQUEST
-            )
+        # Логирование полученного init_data
+        print(f"Received init_data: {init_data}")
 
-        parsed_data = dict(x.split("=") for x in init_data.split("&") if "=" in x)
-        hash_to_check = parsed_data.pop("hash", None)
-        sorted_data_string = "\n".join(
-            f"{k}={v}" for k, v in sorted(parsed_data.items())
-        )
+        # Проверка подписи
+        if not self.check_webapp_signature(TELEGRAM_BOT_TOKEN, init_data):
+            return Response({"error": "Invalid hash. Data has been tampered with."}, status=400)
 
-        # Генерация хеша для проверки
-        secret_key = hashlib.sha256(bot_token.encode()).digest()
-        calculated_hash = hmac.new(
-            secret_key, sorted_data_string.encode(), hashlib.sha256
-        ).hexdigest()
+        # Разбираем параметры init_data
+        params = dict(urllib.parse.parse_qsl(init_data))
 
-        if calculated_hash != hash_to_check:
-            return Response(
-                {"error": "Invalid initData"}, status=status.HTTP_403_FORBIDDEN
-            )
+        # Логирование параметров
+        print(f"Parsed parameters: {params}")
 
-        # Получаем данные пользователя из initData
-        user_data = json.loads(parsed_data.get("user", "{}"))
-        tg_id = user_data.get("id")
-        first_name = user_data.get("first_name", "")
-        last_name = user_data.get("last_name", "")
-        username = user_data.get("username", f"user_{tg_id}")
+        # Декодируем параметр 'user', чтобы избежать экранированных символов
+        user_data = urllib.parse.unquote(params.get('user'))
+        params['user'] = user_data  # Обновляем параметр 'user' в params
 
-        if not tg_id:
-            return Response(
-                {"error": "Invalid Telegram ID"}, status=status.HTTP_400_BAD_REQUEST
-            )
+        # Логирование декодированных данных пользователя
+        print(f"Decoded user data: {user_data}")
 
-        # Получаем или создаем пользователя в базе данных
-        user, created = User.objects.get_or_create(
+        # Декодирование данных из user
+        try:
+            user_data = json.loads(user_data)  # Декодируем user как JSON
+        except json.JSONDecodeError:
+            return Response({"error": "Invalid user data"}, status=400)
+
+        # Логирование данных пользователя
+        print(f"User data: {user_data}")
+
+        tg_id = user_data.get('id')
+        username = user_data.get('username')
+        first_name = user_data.get('first_name')
+        last_name = user_data.get('last_name')
+        avatar_url = user_data.get('photo_url')
+
+        # Создание или получение пользователя
+        user, created = CustomUser.objects.get_or_create(
             tg_id=tg_id,
             defaults={
-                "username": username,
-                "first_name": first_name,
-                "last_name": last_name,
-            },
+                'username': username,
+                'first_name': first_name,
+                'last_name': last_name,
+                'avatar_url': avatar_url
+            }
         )
 
-        tokens = get_tokens_for_user(user)
-        return Response(tokens)
+        serializer = UserSerializer(user)
+
+        # Получение JWT токенов
+        from rest_framework_simplejwt.tokens import RefreshToken
+        refresh = RefreshToken.for_user(user)
+        access_token = str(refresh.access_token)
+
+        return Response({
+            'user': serializer.data,
+            'access_token': access_token,
+            'refresh_token': str(refresh)
+        })
+
+    def check_webapp_signature(self, token: str, init_data: str) -> bool:
+        """
+        Check incoming WebApp init data signature
+
+        Source: https://core.telegram.org/bots/webapps#validating-data-received-via-the-web-app
+
+        :param token:
+        :param init_data:
+        :return:
+        """
+        try:
+            parsed_data = dict(urllib.parse.parse_qsl(init_data))
+        except ValueError:
+            # Init data is not a valid query string
+            return False
+        if "hash" not in parsed_data:
+            # Hash is not present in init data
+            return False
+
+        hash_ = parsed_data.pop('hash')
+        data_check_string = "\n".join(
+            f"{k}={v}" for k, v in sorted(parsed_data.items())
+        )
+        secret_key = hmac.new(
+            key=b"WebAppData", msg=token.encode(), digestmod=hashlib.sha256
+        )
+        calculated_hash = hmac.new(
+            key=secret_key.digest(), msg=data_check_string.encode(), digestmod=hashlib.sha256
+        ).hexdigest()
+        return calculated_hash == hash_
